fix: move shared to libs dir
