+++ libs/portal/data-access/package.json
@@ -7,7 +7,7 @@
     "@app-speed/environments": "*",
     "rxjs": "~7.5.0",
     "@rx-angular/state": "19.0.0",
-    "@app-speed/shared": "0.0.1"
+    "@app-speed/shared-utils": "0.0.1"
   },
   "sideEffects": false
 }
