+++ libs/portal/data-access/src/lib/result.model.ts
@@ -1,3 +1,3 @@
-import { Reports } from '@app-speed/shared';
+import { Reports } from '@app-speed/shared-utils';
 
 export type ResultModel = Reports;
