+++ libs/portal/ui/src/component/results-display/results-display.component.ts
@@ -3,7 +3,7 @@ import { CommonModule } from '@angular/common';
 import { SafeResourceUrl } from '@angular/platform-browser';
 import { LoadingSpinnerComponent } from '@app-speed/portal-ui/loading-spinner';
 import { map, Observable } from 'rxjs';
-import { AuditStatusType } from '@app-speed/shared';
+import { AuditStatusType } from '@app-speed/shared-utils';
 import { RxState } from '@rx-angular/state';
 import { RxIf } from '@rx-angular/template/if';
 import { RxPush } from '@rx-angular/template/push';
