+++ libs/runner/data-access-queue/src/lib/aws-sqs.ts
@@ -10,7 +10,7 @@ import {
 
 import { Message } from '@aws-sdk/client-sqs/dist-types/models/models_0';
 import { AuditQueue } from '@app-speed/runner-interfaces';
-import { AuditRunParams } from '@app-speed/shared';
+import { AuditRunParams } from '@app-speed/shared-utils';
 
 type SQSConfig = {
   region: string;
