+++ libs/runner/interfaces/src/lib/queue.ts
@@ -1,4 +1,4 @@
-import { AuditRunParams } from '@app-speed/shared';
+import { AuditRunParams } from '@app-speed/shared-utils';
 
 export abstract class AuditQueue {
   abstract nextItem(): Promise<AuditRunParams | void>;
