+++ libs/runner/interfaces/src/lib/store.ts
@@ -1,4 +1,4 @@
-import { ResultReports } from '@app-speed/shared';
+import { ResultReports } from '@app-speed/shared-utils';
 
 export interface AuditStore {
   store(auditResults: ResultReports, location?: string): Promise<string>;
