+++ libs/shared/utils/ng-package.json
@@ -1,0 +1,7 @@
+{
+  "$schema": "../../../node_modules/ng-packagr/ng-package.schema.json",
+  "dest": "../../../dist/libs/shared/utils",
+  "lib": {
+    "entryFile": "src/public-api.ts"
+  }
+}
