+++ libs/shared/utils/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "@app-speed/shared",
+  "name": "@app-speed/shared-utils",
   "version": "0.0.1",
   "peerDependencies": {
     "@angular/common": "^15.0.0",
