+++ libs/shared/utils/src/directives/bypass-src.directive.ts
@@ -1,10 +1,10 @@
-import {Directive, ElementRef, Input} from '@angular/core';
+import { Directive, ElementRef, Input } from '@angular/core';
 
 // @TODO - Fix linting rule
 @Directive({
   // eslint-disable-next-line @angular-eslint/directive-selector
   selector: 'iframe[bypassSrc]',
-  standalone: true
+  standalone: true,
 })
 export class BypassSrcDirective {
   constructor(private element: ElementRef) {}
