+++ libs/shared/utils/src/types/types.ts
@@ -1,0 +1,14 @@
+export type Reports = {
+  htmlReportUrl?: string;
+};
+
+export type ResultReports = {
+  jsonReport: string;
+  htmlReport: string;
+};
+
+export type AuditRunParams = {
+  targetUrl: string;
+  requesterId: string;
+  endpoint: string;
+};
