--- packages/shared/ng-package.json
@@ -1,7 +1,0 @@
-{
-  "$schema": "../../node_modules/ng-packagr/ng-package.schema.json",
-  "dest": "../../dist/packages/shared",
-  "lib": {
-    "entryFile": "public-api.ts"
-  }
-}
