--- packages/shared/src/types/types.ts
@@ -1,14 +1,0 @@
-export type Reports = {
-	htmlReportUrl?: string;
-}
-
-export type ResultReports = {
-	jsonReport: string
-	htmlReport: string
-}
-
-export type AuditRunParams = {
-	targetUrl: string;
-	requesterId: string;
-	endpoint: string;
-};
